separated channel components into indiv dir
