+++ react-app/src/components/ChannelCard/ChannelCard.js
