+++ react-app/src/components/ChannelCard/index.js
@@ -1,0 +1,3 @@
+import ChannelCard from "./ChannelCard";
+
+export default ChannelCard;
