+++ react-app/src/components/CreateChannelForm/index.js
@@ -1,0 +1,3 @@
+import CreateChannelForm from "./CreateChannelForm";
+
+export default CreateChannelForm;
