+++ react-app/src/components/DeleteChannelButton/index.js
@@ -1,0 +1,3 @@
+import DeleteChannelButton from "./DeleteChannelButton";
+
+export default DeleteChannelButton;
