+++ react-app/src/components/EditChannelForm/index.js
@@ -1,0 +1,3 @@
+import EditChannelForm from "./EditChannelForm";
+
+export default EditChannelForm;
